+++ test/cli/test_commands.py
@@ -2287,10 +2287,10 @@ def test_parameter_overrides(testdir, cli, schema_url):
 
             @schemathesis.check
             def verify_overrides(response, case):
-                if case.operation.path_parameters.contains("key"):
+                if "key" in case.operation.path_parameters:
                     assert case.path_parameters["key"] == "foo"
                     assert "id" not in (case.query or {}), "`id` is present"
-                if case.operation.query.contains("id"):
+                if "id" in case.operation.query:
                     assert case.query["id"] == "bar"
                     assert "key" not in (case.path_parameters or {}), "`key` is present"
             """